java/src/main/java/me.liyp.leetcode.algorithms/_1_Two_Sum/Solution.java
package me.liyp.leetcode.algorithms._1_Two_Sum;

import java.util.HashMap;

/**
 * @author liyp
 * @since 2018-03-13 00:06:33
 */
public class Solution {

    // O(n^2)
    public int[] twoSum(int[] nums, int target) {
        for (int i = 0; i < nums.length; i++) {
            for (int j = nums.length - 1; j > i; j--) {
                if (nums[i] + nums[j] == target) {
                    return new int[] { i, j };
                }
            }
        }
        throw new IllegalArgumentException();
    }

    // O(n)
    public int[] twoSum_1(int[] nums, int target) {
        HashMap<Integer, Integer> map = new HashMap<Integer, Integer>();
        for (int i = 0; i < nums.length; i++) {
            int complement = target - nums[i];
            if (map.containsKey(complement)) {
                return new int[] {map.get(complement), i};
            }
            map.put(nums[i], i);
        }
        throw new IllegalArgumentException();
    }

}
